python/requirements.txt
audioread==3.1.0
av==17.0.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
decorator==5.2.1
flask==3.1.3
flatbuffers==25.12.19
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.5.3
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.47.0
markupsafe==3.0.3
mpmath==1.3.0
msgpack==1.1.2
numba==0.65.0
numpy==2.4.4
onnxruntime==1.24.4
packaging==26.0
pandas==3.0.2
platformdirs==4.9.4
pooch==1.9.0
protobuf==7.34.1
pyannote-core==6.0.1
pyannote-onnx-extended @ git+https://github.com/samson6460/pyannote-onnx-extended.git@edb7dbc1f6fa5c586064665c016947fc0057a32c
pycparser==3.0
python-dateutil==2.9.0.post0
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
sherpa-onnx==1.12.35
sherpa-onnx-core==1.12.35
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
sympy==1.14.0
threadpoolctl==3.6.0
typing-extensions==4.15.0
urllib3==2.6.3
webrtcvad-wheels==2.0.14
werkzeug==3.1.8
